# Conflicts:
#	cas-server-support-jpa-ticket-registry/src/main/java/org/jasig/cas/ticket/registry/JpaTicketRegistry.java

+++ cas-server-integration-ehcache/src/main/java/org/jasig/cas/ticket/registry/EhCacheTicketRegistry.java
@@ -237,17 +237,17 @@ public final class EhCacheTicketRegistry extends AbstractTicketRegistry implemen
      * @see Cache#getKeysWithExpiryCheck()
      */
     @Override
-    public long sessionCount() {
+    public int sessionCount() {
         return BooleanUtils.toInteger(this.supportRegistryState, this.ticketGrantingTicketsCache
-                .getKeysWithExpiryCheck().size(), (int) super.sessionCount());
+                .getKeysWithExpiryCheck().size(), super.sessionCount());
     }
 
     /**
      * @see Cache#getKeysWithExpiryCheck()
      */
     @Override
-    public long serviceTicketCount() {
+    public int serviceTicketCount() {
         return BooleanUtils.toInteger(this.supportRegistryState, this.serviceTicketsCache.getKeysWithExpiryCheck()
-                .size(), (int) super.serviceTicketCount());
+                .size(), super.serviceTicketCount());
     }
 }
